Updated testdata
