--- testdata/jennies/rawtypes/arrays/TerraformModels/arrays_model_gen.go
@@ -1,14 +1,0 @@
-package provider
-
-import (
-	"github.com/hashicorp/terraform-plugin-framework/types"
-)
-
-
-
-type ArraysSomeStructDataSourceModel struct {
-  
-     FieldAny types.Object `tfsdk:"field_any"`
-  
-}
-
