+++ testdata/jennies/rawtypes/enums/TerraformModels/enums/models_gen.go
@@ -1,8 +1,7 @@
-package provider
+package enums
 
 import (
 	"github.com/hashicorp/terraform-plugin-framework/types"
 )
-
