--- testdata/jennies/rawtypes/field_with_struct_with_defaults/TerraformModels/defaults_model_gen.go
@@ -1,20 +1,0 @@
-package provider
-
-import (
-	"github.com/hashicorp/terraform-plugin-framework/types"
-)
-
-
-
-type DefaultsNestedStructDataSourceModel struct {
-  
-     StringVal types.String `tfsdk:"string_val"`
-  
-     IntVal types.Int64 `tfsdk:"int_val"`
-  
-}
-
-type DefaultsStructDataSourceModel struct {
-  
-}
-
