--- testdata/jennies/rawtypes/intersections/TerraformModels/intersections_model_gen.go
@@ -1,14 +1,0 @@
-package provider
-
-import (
-	"github.com/hashicorp/terraform-plugin-framework/types"
-)
-
-
-
-type IntersectionsSomeStructDataSourceModel struct {
-  
-     FieldBool types.Bool `tfsdk:"field_bool"`
-  
-}
-
