--- testdata/jennies/rawtypes/maps/TerraformModels/maps_model_gen.go
@@ -1,14 +1,0 @@
-package provider
-
-import (
-	"github.com/hashicorp/terraform-plugin-framework/types"
-)
-
-
-
-type MapsSomeStructDataSourceModel struct {
-  
-     FieldAny types.Object `tfsdk:"field_any"`
-  
-}
-
