+++ testdata/jennies/rawtypes/package-with-dashes/TerraformModels/withdashes/models_gen.go
@@ -1,0 +1,18 @@
+package with-dashes
+
+import (
+	"github.com/hashicorp/terraform-plugin-framework/types"
+)
+
+
+type SomeStructDataSourceModel struct {
+     FieldAny types.Object `tfsdk:"field_any"`
+  TemporaryScalarPlaceholder types.Bool // @TODO Remove this once non-scalars are implemented
+}
+
+type StringOrBoolDataSourceModel struct {
+     String types.String `tfsdk:"string"`
+     Bool types.Bool `tfsdk:"bool"`
+  TemporaryScalarPlaceholder types.Bool // @TODO Remove this once non-scalars are implemented
+}
+
