+++ testdata/jennies/rawtypes/scalars/TerraformModels/scalars/models_gen.go
@@ -1,8 +1,7 @@
-package provider
+package scalars
 
 import (
 	"github.com/hashicorp/terraform-plugin-framework/types"
 )
-
