--- testdata/jennies/rawtypes/struct_with_optional_fields/TerraformModels/struct_optional_fields_model_gen.go
@@ -1,20 +1,0 @@
-package provider
-
-import (
-	"github.com/hashicorp/terraform-plugin-framework/types"
-)
-
-
-
-type StructOptionalFieldsSomeStructDataSourceModel struct {
-  
-     FieldString types.String `tfsdk:"field_string"`
-  
-}
-
-type StructOptionalFieldsSomeOtherStructDataSourceModel struct {
-  
-     FieldAny types.Object `tfsdk:"field_any"`
-  
-}
-
