+++ testdata/jennies/rawtypes/time_hint/TerraformModels/time_hint/models_gen.go
@@ -1,0 +1,12 @@
+package time_hint
+
+import (
+	"github.com/hashicorp/terraform-plugin-framework/types"
+)
+
+
+type ObjWithTimeFieldDataSourceModel struct {
+     RegisteredAt types.String `tfsdk:"registered_at"`
+  TemporaryScalarPlaceholder types.Bool // @TODO Remove this once non-scalars are implemented
+}
+
