--- testdata/jennies/rawtypes/time_hint/TerraformModels/time_hint_model_gen.go
@@ -1,14 +1,0 @@
-package provider
-
-import (
-	"github.com/hashicorp/terraform-plugin-framework/types"
-)
-
-
-
-type TimeHintObjWithTimeFieldDataSourceModel struct {
-  
-     RegisteredAt types.String `tfsdk:"registered_at"`
-  
-}
-
