+++ testdata/jennies/rawtypes/variant_dataquery/TerraformModels/variant_dataquery/models_gen.go
@@ -1,0 +1,13 @@
+package variant_dataquery
+
+import (
+	"github.com/hashicorp/terraform-plugin-framework/types"
+)
+
+
+type QueryDataSourceModel struct {
+     Expr types.String `tfsdk:"expr"`
+     Instant types.Bool `tfsdk:"instant"`
+  TemporaryScalarPlaceholder types.Bool // @TODO Remove this once non-scalars are implemented
+}
+
