--- testdata/jennies/rawtypes/variant_dataquery/TerraformModels/variant_dataquery_model_gen.go
@@ -1,16 +1,0 @@
-package provider
-
-import (
-	"github.com/hashicorp/terraform-plugin-framework/types"
-)
-
-
-
-type VariantDataqueryQueryDataSourceModel struct {
-  
-     Expr types.String `tfsdk:"expr"`
-  
-     Instant types.Bool `tfsdk:"instant"`
-  
-}
-
