--- testdata/jennies/rawtypes/variant_panelcfg_full/TerraformModels/variant_panelcfg_full_model_gen.go
@@ -1,20 +1,0 @@
-package provider
-
-import (
-	"github.com/hashicorp/terraform-plugin-framework/types"
-)
-
-
-
-type VariantPanelcfgFullOptionsDataSourceModel struct {
-  
-     TimeseriesOption types.String `tfsdk:"timeseries_option"`
-  
-}
-
-type VariantPanelcfgFullFieldConfigDataSourceModel struct {
-  
-     TimeseriesFieldConfigOption types.String `tfsdk:"timeseries_field_config_option"`
-  
-}
-
