+++ testdata/jennies/rawtypes/variant_panelcfg_only_options/TerraformModels/variant_panelcfg_only_options/models_gen.go
@@ -1,0 +1,12 @@
+package variant_panelcfg_only_options
+
+import (
+	"github.com/hashicorp/terraform-plugin-framework/types"
+)
+
+
+type OptionsDataSourceModel struct {
+     Content types.String `tfsdk:"content"`
+  TemporaryScalarPlaceholder types.Bool // @TODO Remove this once non-scalars are implemented
+}
+
