--- testdata/jennies/rawtypes/variant_panelcfg_only_options/TerraformModels/variant_panelcfg_only_options_model_gen.go
@@ -1,14 +1,0 @@
-package provider
-
-import (
-	"github.com/hashicorp/terraform-plugin-framework/types"
-)
-
-
-
-type VariantPanelcfgOnlyOptionsOptionsDataSourceModel struct {
-  
-     Content types.String `tfsdk:"content"`
-  
-}
-
